split ui.py, more accurate remaining time calculation, better progressbar
