+++ soundconverter/interface/ui.py
@@ -20,91 +20,24 @@
 # USA
 
 import os
-import time
 import sys
-import urllib.request
-import urllib.parse
-import urllib.error
 from gettext import gettext as _
-from gettext import ngettext
 
-from gi.repository import GObject, Gtk, Gio, Gdk, GLib, Pango
+from gi.repository import Gtk, GLib
 
-from soundconverter.util.fileoperations import filename_to_uri, \
-    beautify_uri, unquote_filename, vfs_walk
-from soundconverter.util.soundfile import SoundFile
-from soundconverter.util.settings import get_gio_settings, settings
-from soundconverter.util.formats import get_quality, \
-    get_bitrate_from_settings, get_file_extension
+from soundconverter.util.fileoperations import filename_to_uri
+from soundconverter.util.settings import settings
 from soundconverter.util.namegenerator import TargetNameGenerator, \
-    subfolder_patterns, basename_patterns, locale_patterns_dict, \
     filepattern
 from soundconverter.util.taskqueue import TaskQueue
 from soundconverter.util.logger import logger
-from soundconverter.gstreamer.discoverer import add_discoverers
-from soundconverter.gstreamer.converter import Converter, available_elements
-from soundconverter.util.error import show_error, set_error_handler
-from soundconverter.interface.notify import notification
+from soundconverter.gstreamer.converter import Converter
+from soundconverter.util.error import set_error_handler
 from soundconverter.util.formatting import format_time
-
-
-# Names of columns in the file list
-MODEL = [
-    GObject.TYPE_STRING,  # visible filename
-    GObject.TYPE_PYOBJECT,  # soundfile
-    GObject.TYPE_FLOAT,  # progress
-    GObject.TYPE_STRING,  # status
-    GObject.TYPE_STRING,  # complete filename
-    GObject.TYPE_FLOAT  # duration of the audio
-]
-
-COLUMNS = ['filename']
-
-# VISIBLE_COLUMNS = ['filename']
-# ALL_COLUMNS = VISIBLE_COLUMNS + ['META']
-
-
-encoders = (
-    ('audio/x-vorbis', 'vorbisenc'),
-    ('audio/mpeg', 'lamemp3enc'),
-    ('audio/x-flac', 'flacenc'),
-    ('audio/x-wav', 'wavenc'),
-    ('audio/x-m4a', 'fdkaacenc,faac,avenc_aac'),
-    ('audio/ogg; codecs=opus', 'opusenc'),
-)  # must be in same order as the output_mime_type GtkComboBox
-
-
-def idle(func):
-    def callback(*args, **kwargs):
-        GLib.idle_add(func, *args, **kwargs)
-    return callback
-
-
-def gtk_iteration(blocking=False):
-    """Keeps the UI and event loops for gst going.
-
-    Paramters
-    ---------
-    blocking : bool
-        If True, will call main_iteration even if no events are pending,
-        which will wait until an event is available.
-    """
-    if blocking:
-        while True:
-            Gtk.main_iteration()
-            if not Gtk.events_pending():
-                break
-    else:
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-
-def gtk_sleep(duration):
-    """Sleep while keeping the GUI responsive."""
-    start = time.time()
-    while time.time() < start + duration:
-        time.sleep(0.01)
-        gtk_iteration()
+from soundconverter.interface.filelist import FileList
+from soundconverter.interface.gladewindow import GladeWindow
+from soundconverter.interface.preferences import PreferencesDialog
+from soundconverter.interface.mainloop import gtk_sleep, gtk_iteration
 
 
 class ErrorDialog:
@@ -125,863 +58,79 @@ class ErrorDialog:
         self.dialog.hide()
 
 
-class FileList:
-    """List of files added by the user."""
+class ProgressBar:
+    """Wrapper for the progressbar to enable smoothing.
 
-    # List of MIME types which we accept for drops.
-    drop_mime_types = ['text/uri-list', 'text/plain', 'STRING']
-
-    def __init__(self, window, builder):
-        self.window = window
-        self.discoverers = None
-        self.filelist = set()
-
-        self.model = Gtk.ListStore(*MODEL)
-
-        self.widget = builder.get_object('filelist')
-        self.widget.props.fixed_height_mode = True
-
-        # sort the longest audio to the front, because converting those
-        # right at the start make sure that no process ends up converting
-        # a single large file while all other tasks are done and therefore
-        # all other cpu cores are idle
-        self.sortedmodel = Gtk.TreeModelSort(model=self.model)
-        self.widget.set_model(self.sortedmodel)
-        self.sortedmodel.set_sort_column_id(4, Gtk.SortType.DESCENDING)
-        self.widget.get_selection().set_mode(Gtk.SelectionMode.MULTIPLE)
-
-        self.widget.drag_dest_set(
-            Gtk.DestDefaults.ALL, [], Gdk.DragAction.COPY
-        )
-        targets = [
-            (accepted, 0, i) for i, accepted
-            in enumerate(self.drop_mime_types)
-        ]
-        self.widget.drag_dest_set_target_list(targets)
-
-        self.widget.connect('drag-data-received', self.drag_data_received)
-
-        renderer = Gtk.CellRendererProgress()
-        column = Gtk.TreeViewColumn(
-            'progress',
-            renderer,
-            value=2,
-            text=3,
-        )
-        column.props.sizing = Gtk.TreeViewColumnSizing.FIXED
-        self.widget.append_column(column)
-        self.progress_column = column
-        self.progress_column.set_visible(False)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property('ellipsize', Pango.EllipsizeMode.MIDDLE)
-        column = Gtk.TreeViewColumn(
-            'Filename',
-            renderer,
-            markup=0,
-        )
-        column.props.sizing = Gtk.TreeViewColumnSizing.FIXED
-        column.set_expand(True)
-        self.widget.append_column(column)
-
-        self.window.progressbarstatus.hide()
-
-        self.invalid_files_list = []
-        self.good_uris = []
-
-    def drag_data_received(self, widget, context, x, y, selection, mime_id, time):
-        widget.stop_emission('drag-data-received')
-        if 0 <= mime_id < len(self.drop_mime_types):
-            text = selection.get_data().decode('utf-8')
-            uris = [uri.strip() for uri in text.split('\n')]
-            self.add_uris(uris)
-            context.finish(True, False, time)
-
-    def get_files(self):
-        """Return all valid SoundFile objects."""
-        return [i[1] for i in self.sortedmodel]
-
-    @idle
-    def add_uris(self, uris, base=None, extensions=None):
-        """Add URIs that should be converted to the list in the GTK interface.
-
-        uris is a list of string URIs, which are absolute paths
-        starting with 'file://'
-
-        extensions is a list of strings like ['.ogg', '.oga'],
-        in which case only files of this type are added to the
-        list. This can be useful when files of multiple types
-        are inside a directory and only some of them should be
-        converted. Default:None which accepts all types.
-        """
-        if len(uris) == 0:
-            return
-
-        start_t = time.time()
-        files = []
-        self.window.set_status(_('Scanning files…'))
-        # for whichever reason, that set_status needs some more iterations
-        # to show up:
-        gtk_iteration(True)
-        self.window.progressbarstatus.show()
-        self.window.progressbarstatus.set_fraction(0)
-
-        for uri in uris:
-            gtk_iteration()
-            if not uri:
-                continue
-            if uri.startswith('cdda:'):
-                show_error(
-                    'Cannot read from Audio CD.',
-                    'Use SoundJuicer Audio CD Extractor instead.'
-                )
-                return
-            info = Gio.file_parse_name(uri).query_file_type(
-                Gio.FileMonitorFlags.NONE, None
-            )
-            if info == Gio.FileType.DIRECTORY:
-                logger.info('walking: \'{}\''.format(uri))
-                if len(uris) == 1:
-                    # if only one folder is passed to the function,
-                    # use its parent as base path.
-                    base = os.path.dirname(uri)
-
-                # get a list of all the files as URIs in
-                # that directory and its subdirectories
-                filelist = vfs_walk(uri)
-
-                accepted = []
-                if extensions:
-                    for filename in filelist:
-                        for extension in extensions:
-                            if filename.lower().endswith(extension):
-                                accepted.append(filename)
-                    filelist = accepted
-                files.extend(filelist)
-            else:
-                files.append(uri)
-
-        files = [f for f in files if not f.endswith('~SC~')]
-
-        if len(files) == 0:
-            show_error('No files found!', '')
-
-        if not base:
-            base = os.path.commonprefix(files)
-            if base and not base.endswith('/'):
-                # we want a common folder
-                base = base[0:base.rfind('/')]
-                base += '/'
-        else:
-            base += '/'
-
-        scan_t = time.time()
-        logger.info('analysing file integrity')
-
-        # self.good_uris will be populated
-        # by the discoverer.
-        # It is a list of uris and only contains those files
-        # that can be handled by gstreamer
-        self.good_uris = []
-
-        self.discoverers = TaskQueue()
-        sound_files = []
-        for filename in files:
-            sound_file = SoundFile(filename, base)
-            sound_files.append(sound_file)
-
-        add_discoverers(self.discoverers, sound_files)
-
-        self.discoverers.set_on_queue_finished(self.discoverer_queue_ended)
-        self.discoverers.run()
-
-        self.window.set_status('{}'.format(_('Adding Files…')))
-        logger.info('adding: {} files'.format(len(files)))
-
-        # show progress and enable GTK main loop iterations
-        # so that the ui stays responsive
-        self.window.progressbarstatus.set_text('0/{}'.format(len(files)))
-        self.window.progressbarstatus.set_show_text(True)
-
-        while self.discoverers.running:
-            progress = self.discoverers.get_progress()[0]
-            if progress:
-                completed = int(progress * len(files))
-                self.window.progressbarstatus.set_fraction(progress)
-                self.window.progressbarstatus.set_text(
-                    '{}/{}'.format(completed, len(files))
-                )
-            gtk_iteration()
-        logger.info('Discovered {} audiofiles in {} s'.format(
-            len(files), round(self.discoverers.get_duration(), 1)
-        ))
-
-        self.window.progressbarstatus.set_show_text(False)
-
-        # see if one of the files with an audio extension
-        # was not readable.
-        known_audio_types = [
-            '.flac', '.mp3', '.aac',
-            '.m4a', '.mpeg', '.opus', '.vorbis', '.ogg', '.wav'
-        ]
-
-        # invalid_files is the number of files that are not
-        # added to the list in the current function call
-        invalid_files = 0
-        # out of those files, that many have an audio file extension
-        broken_audiofiles = 0
-
-        sound_files = []
-        for discoverer in self.discoverers.all_tasks:
-            sound_files += discoverer.sound_files
-
-        for sound_file in sound_files:
-            # create a list of human readable file paths
-            # that were not added to the list
-            if not sound_file.readable:
-                filename = sound_file.filename
-
-                extension = os.path.splitext(filename)[1].lower()
-                if extension in known_audio_types:
-                    broken_audiofiles += 1
-
-                subfolders = sound_file.subfolders
-                relative_path = os.path.join(subfolders, filename)
-
-                self.invalid_files_list.append(relative_path)
-                invalid_files += 1
-                continue
-            if sound_file.uri in self.filelist:
-                logger.info('file already present: \'{}\''.format(
-                    sound_file.uri
-                ))
-                continue
-            self.append_file(sound_file)
-
-        if invalid_files > 0:
-            self.window.invalid_files_button.set_visible(True)
-            if len(files) == invalid_files == 1:
-                # case 1: the single file that should be added is not supported
-                show_error(
-                    _('The specified file is not supported!'),
-                    _('Either because it is broken or not an audio file.')
-                )
-
-            elif len(files) == invalid_files:
-                # case 2: all files that should be added cannot be added
-                show_error(
-                    _('All {} specified files are not supported!').format(
-                        len(files)
-                    ),
-                    _('Either because they are broken or not audio files.')
-                )
-
-            else:
-                # case 3: some files could not be added (that can already be
-                # because there is a single picture in a folder of hundreds
-                # of sound files). Show an error if this skipped file has a
-                # soundfile extension, otherwise don't bother the user.
-                logger.info(
-                    '{} of {} files were not added to the list'.format(
-                        invalid_files, len(files)
-                    )
-                )
-                if broken_audiofiles > 0:
-                    show_error(
-                        ngettext(
-                            'One audio file could not be read by GStreamer!',
-                            '{} audio files could not be read by GStreamer!',
-                            broken_audiofiles
-                        ).format(broken_audiofiles),
-                        _(
-                            'Check "Invalid Files" in the menu for more'
-                            'information.'
-                        )
-                    )
-        else:
-            # case 4: all files were successfully added. No error message
-            pass
-
-        self.window.set_status()
-        self.window.progressbarstatus.hide()
-        end_t = time.time()
-        logger.debug(
-            'Added %d files in %.2fs (scan %.2fs, add %.2fs)' % (
-                len(files), end_t - start_t, scan_t - start_t, end_t - scan_t
-            )
-        )
-
-    def discoverer_queue_ended(self, queue):
-        # all tasks done
-        self.window.set_sensitive()
-        self.window.conversion_ended()
-
-        total_time = queue.get_duration()
-        msg = _('Tasks done in %s') % format_time(total_time)
-
-        errors = [
-            task.error for task in queue.done
-            if task.error is not None
-        ]
-        if len(errors) > 0:
-            msg += ', {} error(s)'.format(len(errors))
-
-        self.window.set_status(msg)
-        if not self.window.is_active():
-            notification(msg)
-
-        readable = []
-        for discoverer in self.discoverers.all_tasks:
-            for sound_file in discoverer.sound_files:
-                if sound_file.readable:
-                    readable.append(sound_file)
-
-        self.good_uris = [sound_file.uri for sound_file in readable]
-        self.window.set_status()
-        self.window.progressbarstatus.hide()
-
-    def cancel(self):
-        if self.discoverers is not None:
-            self.discoverers.cancel()
-
-    def format_cell(self, sound_file):
-        """Take a SoundFile and return a human readable path to it."""
-        return GLib.markup_escape_text(unquote_filename(sound_file.filename))
-
-    def set_row_progress(self, number, progress):
-        """Update the progress bar of a single row/file."""
-        self.progress_column.set_visible(True)
-        if self.model[number][2] == 1.0:
-            return
-
-        self.model[number][2] = progress * 100.0
-
-    def hide_row_progress(self):
-        self.progress_column.set_visible(False)
-
-    def append_file(self, sound_file):
-        """Add a valid SoundFile object to the list of files in the GUI.
-
-        Parameters
-        ----------
-        sound_file : SoundFile
-            This soundfile is expected to be readable by gstreamer
-        """
-        self.model.append([
-            self.format_cell(sound_file),
-            sound_file,
-            0.0,
-            '',
-            sound_file.uri,
-            sound_file.duration
-        ])
-        self.filelist.add(sound_file.uri)
-        sound_file.filelist_row = len(self.model) - 1
-
-    def remove(self, iterator):
-        uri = self.model.get(iterator, 1)[0].uri
-        self.filelist.remove(uri)
-        self.model.remove(iterator)
-
-    def is_nonempty(self):
-        try:
-            self.model.get_iter((0,))
-        except ValueError:
-            return False
-        return True
-
-
-class GladeWindow(object):
-
-    callbacks = {}
-    builder = None
-
-    def __init__(self, builder):
-        """Init GladeWindow, store the objects's potential callbacks for later.
-
-        You have to call connect_signals() when all descendants are ready.
-        """
-        GladeWindow.builder = builder
-        GladeWindow.callbacks.update(dict(
-            [[x, getattr(self, x)] for x in dir(self) if x.startswith('on_')]
-        ))
+    Because changes only arrive every second which makes the progress jump
+    for a small number of files.
+    """
+    def __init__(self, progressbar, period=1):
+        """Initialize the fraction object without doing anything yet."""
+        self.progressbar = progressbar
+        self.steps = 0
+        self.period = period
+        self.current_fraction = 0
+        self.fraction_target = 0
+        self.step = 0
+        self.timeout_id = None
+        self.progressbar.set_fraction(0)
 
     def __getattr__(self, attribute):
-        """Allow direct use of window widget."""
-        widget = GladeWindow.builder.get_object(attribute)
-        if widget is None:
-            raise AttributeError('Widget \'{}\' not found'.format(attribute))
-        self.__dict__[attribute] = widget  # cache result
-        return widget
-
-    @staticmethod
-    def connect_signals():
-        """Connect all GladeWindow objects to theirs respective signals."""
-        GladeWindow.builder.connect_signals(GladeWindow.callbacks)
-
-
-class PreferencesDialog(GladeWindow):
-    sensitive_names = [
-        'vorbis_quality', 'choose_folder', 'create_subfolders',
-        'subfolder_pattern', 'jobs_spinbutton', 'resample_hbox',
-        'force_mono'
-    ]
-
-    def __init__(self, builder, parent):
-        self.settings = get_gio_settings()
-        GladeWindow.__init__(self, builder)
-
-        # encoders should be in the same order as in the glade file
-        for i, encoder in enumerate(encoders):
-            extension = get_file_extension(encoder[0])
-            ui_row_text = self.liststore8[i][0]
-            # the extension should be part of the row with the correct index
-            assert extension.lower() in ui_row_text.lower()
-
-        self.dialog = builder.get_object('prefsdialog')
-        self.dialog.set_transient_for(parent)
-        self.example = builder.get_object('example_filename')
-        self.force_mono = builder.get_object('force_mono')
-
-        self.target_bitrate = None
-
-        self.sensitive_widgets = {}
-        for name in self.sensitive_names:
-            self.sensitive_widgets[name] = builder.get_object(name)
-            assert self.sensitive_widgets[name] is not None
-
-        self.encoders = None
-        self.present_mime_types = []
-
-        self.set_widget_initial_values()
-        self.set_sensitive()
-
-        tip = [_('Available patterns:')]
-        for k in sorted(locale_patterns_dict.values()):
-            tip.append(k)
-        self.custom_filename.set_tooltip_text('\n'.join(tip))
-
-    def set_widget_initial_values(self):
-        self.quality_tabs.set_show_tabs(False)
-
-        if self.settings.get_boolean('same-folder-as-input'):
-            widget = self.same_folder_as_input
-        else:
-            widget = self.into_selected_folder
-        widget.set_active(True)
-
-        self.target_folder_chooser = Gtk.FileChooserDialog(
-            title=_('Add Folder…'),
-            transient_for=self.dialog,
-            action=Gtk.FileChooserAction.SELECT_FOLDER
-        )
-
-        self.target_folder_chooser.add_button(
-            Gtk.STOCK_CANCEL, Gtk.ResponseType.CANCEL
-        )
-        self.target_folder_chooser.add_button(
-            Gtk.STOCK_OPEN, Gtk.ResponseType.OK
-        )
-
-        self.target_folder_chooser.set_select_multiple(False)
-        self.target_folder_chooser.set_local_only(False)
-
-        uri = filename_to_uri(urllib.parse.quote(
-            self.settings.get_string('selected-folder'), safe='/:@'
-        ))
-        self.target_folder_chooser.set_uri(uri)
-        self.update_selected_folder()
-
-        widget = self.create_subfolders
-        widget.set_active(self.settings.get_boolean('create-subfolders'))
-
-        widget = self.subfolder_pattern
-        active = self.settings.get_int('subfolder-pattern-index')
-        model = widget.get_model()
-        model.clear()
-        for pattern, desc in subfolder_patterns:
-            i = model.append()
-            model.set(i, 0, desc)
-        widget.set_active(active)
-
-        if self.settings.get_boolean('replace-messy-chars'):
-            widget = self.replace_messy_chars
-            widget.set_active(True)
-
-        if self.settings.get_boolean('delete-original'):
-            self.delete_original.set_active(True)
-
-        current_mime_type = self.settings.get_string('output-mime-type')
-
-        # deactivate output if encoder plugin is not present
-        widget = self.output_mime_type
-        model = widget.get_model()
-        assert len(model) == len(encoders), 'model:{} widgets:{}'.format(
-            len(model), len(encoders)
-        )
-
-        i = 0
-        model = self.output_mime_type.get_model()
-        for mime, encoder_name in encoders:
-            # valid default output?
-            encoder_present = any(
-                e in available_elements for e in encoder_name.split(',')
-            )
-            if encoder_present:
-                # add to supported outputs
-                self.present_mime_types.append(mime)
-                i += 1
-            else:
-                logger.error(
-                    '{} {} is not supported, a gstreamer plugins package '
-                    'is possibly missing.'.format(mime, encoder_name)
-                )
-                del model[i]
-
-        for i, mime in enumerate(self.present_mime_types):
-            if current_mime_type == mime:
-                widget.set_active(i)
-        self.change_mime_type(current_mime_type)
-
-        # display information about mp3 encoding
-        if 'lamemp3enc' not in available_elements:
-            widget = self.lame_absent
-            widget.show()
-
-        widget = self.vorbis_quality
-        quality = self.settings.get_double('vorbis-quality')
-        quality_setting = get_quality('audio/x-vorbis', quality, reverse=True)
-        widget.set_active(-1)
-        self.vorbis_quality.set_active(quality_setting)
-        if self.settings.get_boolean('vorbis-oga-extension'):
-            self.vorbis_oga_extension.set_active(True)
-
-        widget = self.aac_quality
-        quality = self.settings.get_int('aac-quality')
-        quality_setting = get_quality('audio/x-m4a', quality, reverse=True)
-        widget.set_active(quality_setting)
-
-        widget = self.opus_quality
-        quality = self.settings.get_int('opus-bitrate')
-        quality_setting = get_quality('audio/ogg; codecs=opus', quality, reverse=True)
-        widget.set_active(quality_setting)
-
-        widget = self.flac_compression
-        quality = self.settings.get_int('flac-compression')
-        quality_setting = get_quality('audio/x-flac', quality, reverse=True)
-        widget.set_active(quality_setting)
-
-        widget = self.wav_sample_width
-        quality = self.settings.get_int('wav-sample-width')
-        quality_setting = get_quality('audio/x-wav', quality, reverse=True)
-        widget.set_active(quality_setting)
-
-        self.mp3_quality = self.mp3_quality
-        self.mp3_mode = self.mp3_mode
-
-        mode = self.settings.get_string('mp3-mode')
-        self.change_mp3_mode(mode)
-
-        widget = self.basename_pattern
-        active = self.settings.get_int('name-pattern-index')
-        model = widget.get_model()
-        model.clear()
-        for pattern, desc in basename_patterns:
-            iter = model.append()
-            model.set(iter, 0, desc)
-        widget.set_active(active)
-
-        self.custom_filename.set_text(
-            self.settings.get_string('custom-filename-pattern')
-        )
-        if self.basename_pattern.get_active() == len(basename_patterns)-1:
-            self.custom_filename_box.set_sensitive(True)
-        else:
-            self.custom_filename_box.set_sensitive(False)
-
-        output_resample = self.settings.get_boolean('output-resample')
-        self.resample_toggle.set_active(output_resample)
-
-        cell = Gtk.CellRendererText()
-        self.resample_rate.pack_start(cell, True)
-        self.resample_rate.add_attribute(cell, 'text', 0)
-        rates = [8000, 11025, 16000, 22050, 32000, 44100, 48000, 96000, 128000]
-        rate = self.settings.get_int('resample-rate')
-        try:
-            idx = rates.index(rate)
-        except ValueError:
-            idx = -1
-        self.resample_rate.set_active(idx)
-
-        self.force_mono.set_active(self.settings.get_boolean('force-mono'))
-
-        self.jobs.set_active(self.settings.get_boolean('limit-jobs'))
-        self.jobs_spinbutton.set_value(self.settings.get_int('number-of-jobs'))
-
-        self.update_example()
-
-    def update_selected_folder(self):
-        self.into_selected_folder.set_use_underline(False)
-        self.into_selected_folder.set_label(
-            _('Into folder %s') %
-            beautify_uri(self.settings.get_string('selected-folder'))
-        )
-
-    def update_example(self):
-        """Refresh the example in the settings dialog."""
-        sound_file = SoundFile('file:///foo/bar.flac')
-        sound_file.tags.update({
-            'track-number': 1,
-            'track-count': 99,
-            'album-disc-number': 2,
-            'album-disc-count': 9
-        })
-        sound_file.tags.update(locale_patterns_dict)
-
-        generator = TargetNameGenerator()
-        generator.replace_messy_chars = False
-
-        example_path = GLib.markup_escape_text(
-            generator.generate_target_uri(sound_file, for_display=True)
-        )
-        position = 0
-        replaces = []
-
-        while True:
-            beginning = example_path.find('{', position)
-            if beginning == -1:
-                break
-            end = example_path.find('}', beginning)
-
-            tag = example_path[beginning:end+1]
-            available_tags = [
-                v.lower() for v in list(locale_patterns_dict.values())
-            ]
-            if tag.lower() in available_tags:
-                bold_tag = tag.replace(
-                    '{', '<b>{'
-                ).replace(
-                    '}', '}</b>'
-                )
-                replaces.append([tag, bold_tag])
-            else:
-                red_tag = tag.replace(
-                    '{', '<span foreground=\'red\'><i>{'
-                ).replace(
-                    '}', '}</i></span>'
-                )
-                replaces.append([tag, red_tag])
-            position = beginning + 1
-
-        for tag, formatted in replaces:
-            example_path = example_path.replace(tag, formatted)
-
-        self.example.set_markup(example_path)
-
-        markup = '<small>{}</small>'.format(
-            _('Target bitrate: %s') % get_bitrate_from_settings()
-        )
-        self.approx_bitrate.set_markup(markup)
-
-    def set_sensitive(self):
-        for widget in list(self.sensitive_widgets.values()):
-            widget.set_sensitive(False)
-
-        same_folder = self.settings.get_boolean('same-folder-as-input')
-        for name in ['choose_folder', 'create_subfolders',
-                     'subfolder_pattern']:
-            self.sensitive_widgets[name].set_sensitive(not same_folder)
-
-        self.sensitive_widgets['vorbis_quality'].set_sensitive(
-            self.settings.get_string('output-mime-type') == 'audio/x-vorbis')
-
-        self.sensitive_widgets['jobs_spinbutton'].set_sensitive(
-            self.settings.get_boolean('limit-jobs'))
-
-        self.sensitive_widgets['resample_hbox'].set_sensitive(True)
-        self.sensitive_widgets['force_mono'].set_sensitive(True)
-
-    def run(self):
-        self.dialog.run()
-        self.dialog.hide()
-
-    def on_delete_original_toggled(self, button):
-        self.settings.set_boolean('delete-original', button.get_active())
-
-    def on_same_folder_as_input_toggled(self, button):
-        self.settings.set_boolean('same-folder-as-input', True)
-        self.set_sensitive()
-        self.update_example()
-
-    def on_into_selected_folder_toggled(self, button):
-        self.settings.set_boolean('same-folder-as-input', False)
-        self.set_sensitive()
-        self.update_example()
-
-    def on_choose_folder_clicked(self, button):
-        ret = self.target_folder_chooser.run()
-        folder = self.target_folder_chooser.get_uri()
-        self.target_folder_chooser.hide()
-        if ret == Gtk.ResponseType.OK:
-            if folder:
-                folder = urllib.parse.unquote(folder)
-                self.settings.set_string('selected-folder', folder)
-                self.update_selected_folder()
-                self.update_example()
-
-    def on_create_subfolders_toggled(self, button):
-        self.settings.set_boolean('create-subfolders', button.get_active())
-        self.update_example()
-
-    def on_subfolder_pattern_changed(self, combobox):
-        self.settings.set_int('subfolder-pattern-index', combobox.get_active())
-        self.update_example()
-
-    def on_basename_pattern_changed(self, combobox):
-        self.settings.set_int('name-pattern-index', combobox.get_active())
-        if combobox.get_active() == len(basename_patterns)-1:
-            self.custom_filename_box.set_sensitive(True)
-        else:
-            self.custom_filename_box.set_sensitive(False)
-        self.update_example()
-
-    def on_custom_filename_changed(self, entry):
-        self.settings.set_string('custom-filename-pattern', entry.get_text())
-        self.update_example()
-
-    def on_replace_messy_chars_toggled(self, button):
-        self.settings.set_boolean('replace-messy-chars', button.get_active())
-
-    def change_mime_type(self, mime_type):
-        """Show the correct quality tab based on the selected format."""
-        self.settings.set_string('output-mime-type', mime_type)
-        self.set_sensitive()
-        self.update_example()
-        tabs = {
-            'audio/x-vorbis': 0,
-            'audio/mpeg': 1,
-            'audio/x-flac': 2,
-            'audio/x-wav': 3,
-            'audio/x-m4a': 4,
-            'audio/ogg; codecs=opus': 5,
-        }
-        self.quality_tabs.set_current_page(tabs[mime_type])
-
-    def on_output_mime_type_changed(self, combo):
-        """Called when the format is changed on the UI."""
-        format_index = combo.get_active()
-        mime = encoders[format_index][0]
-        if mime in self.present_mime_types:
-            self.change_mime_type(mime)
-
-    def on_output_mime_type_ogg_vorbis_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/x-vorbis')
-
-    def on_output_mime_type_flac_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/x-flac')
-
-    def on_output_mime_type_wav_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/x-wav')
-
-    def on_output_mime_type_mp3_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/mpeg')
-
-    def on_output_mime_type_aac_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/x-m4a')
-
-    def on_output_mime_type_opus_toggled(self, button):
-        if button.get_active():
-            self.change_mime_type('audio/ogg; codecs=opus')
-
-    def on_vorbis_quality_changed(self, combobox):
-        if combobox.get_active() == -1:
-            return  # just de-selectionning
-        fquality = get_quality('audio/x-vorbis', combobox.get_active())
-        self.settings.set_double('vorbis-quality', fquality)
-        self.update_example()
-
-    def on_vorbis_oga_extension_toggled(self, toggle):
-        self.settings.set_boolean('vorbis-oga-extension', toggle.get_active())
-        self.update_example()
-
-    def on_aac_quality_changed(self, combobox):
-        quality = get_quality('audio/x-m4a', combobox.get_active())
-        self.settings.set_int('aac-quality', quality)
-        self.update_example()
-
-    def on_opus_quality_changed(self, combobox):
-        quality = get_quality('audio/ogg; codecs=opus', combobox.get_active())
-        self.settings.set_int('opus-bitrate', quality)
-        self.update_example()
-
-    def on_wav_sample_width_changed(self, combobox):
-        quality = get_quality('audio/x-wav', combobox.get_active())
-        self.settings.set_int('wav-sample-width', quality)
-        self.update_example()
-
-    def on_flac_compression_changed(self, combobox):
-        quality = get_quality('audio/x-flac', combobox.get_active())
-        self.settings.set_int('flac-compression', quality)
-        self.update_example()
-
-    def on_force_mono_toggle(self, button):
-        self.settings.set_boolean('force-mono', button.get_active())
-        self.update_example()
-
-    def change_mp3_mode(self, mode):
-        keys = {'cbr': 0, 'abr': 1, 'vbr': 2}
-        self.mp3_mode.set_active(keys[mode])
-
-        keys = {
-            'cbr': 'mp3-cbr-quality',
-            'abr': 'mp3-abr-quality',
-            'vbr': 'mp3-vbr-quality',
-        }
-        quality = self.settings.get_int(keys[mode])
-
-        index = get_quality('audio/mpeg', quality, mode, reverse=True)
-        self.mp3_quality.set_active(index)
-        self.update_example()
-
-    def on_mp3_mode_changed(self, combobox):
-        mode = ('cbr', 'abr', 'vbr')[combobox.get_active()]
-        self.settings.set_string('mp3-mode', mode)
-        self.change_mp3_mode(mode)
-
-    def on_mp3_quality_changed(self, combobox):
-        keys = {
-            'cbr': 'mp3-cbr-quality',
-            'abr': 'mp3-abr-quality',
-            'vbr': 'mp3-vbr-quality'
-        }
-        mode = self.settings.get_string('mp3-mode')
-
-        bitrate = get_quality('audio/mpeg', combobox.get_active(), mode)
-        self.settings.set_int(keys[mode], bitrate)
-        self.update_example()
-
-    def on_resample_rate_changed(self, combobox):
-        selected = combobox.get_active()
-        rates = [8000, 11025, 16000, 22050, 32000, 44100, 48000, 96000, 128000]
-        self.settings.set_int('resample-rate', rates[selected])
-        self.update_example()
-
-    def on_resample_toggle(self, rstoggle):
-        self.settings.set_boolean('output-resample', rstoggle.get_active())
-        self.resample_rate.set_sensitive(rstoggle.get_active())
-        self.update_example()
-
-    def on_jobs_toggled(self, jtoggle):
-        self.settings.set_boolean('limit-jobs', jtoggle.get_active())
-        self.jobs_spinbutton.set_sensitive(jtoggle.get_active())
-
-    def on_jobs_spinbutton_value_changed(self, jspinbutton):
-        self.settings.set_int('number-of-jobs', int(jspinbutton.get_value()))
+        """For all other attributes, try to get them from the widget."""
+        return getattr(self.progressbar, attribute)
+
+    def set_fraction(self, fraction):
+        """Set a fraction that will be shown after interpolating to it."""
+        if fraction in [0, 1]:
+            self.set_current(fraction)
+            return
+
+        self.fraction_target = fraction
+        difference = self.fraction_target - self.current_fraction
+
+        # not more steps than the progressbar can resolute
+        self.steps = round(self.progressbar.get_allocated_width() * difference)
+
+        # don't make an interval if not needed
+        if self.steps <= 1:
+            self.set_current(fraction)
+            return
+
+        # restart the interpolation
+        self.step = 0
+        interval = 1000 * self.period / self.steps
+        if self.timeout_id is not None:
+            GLib.Source.remove(self.timeout_id)
+        self.timeout_id = GLib.timeout_add(interval, self.interpolate)
+
+    def get_fraction(self):
+        return self.current_fraction
+
+    def set_current(self, new_fraction):
+        """Set the value immediatelly without interpolating it."""
+        if new_fraction != self.current_fraction:
+            self.progressbar.set_fraction(new_fraction)
+            self.current_fraction = new_fraction
+
+    def interpolate(self):
+        """Take a small step from the previous fraction to the target."""
+        if self.fraction_target in [0, 1]:
+            self.set_current(self.fraction_target)
+            self.timeout_id = None
+            return False
+
+        difference = self.fraction_target - self.current_fraction
+        change = difference / max(1, self.steps - self.step)
+        self.step += 1
+        new_fraction = self.current_fraction + change
+        self.set_current(new_fraction)
+
+        if self.step >= self.steps:
+            # wait before starting the interpolation timeout again
+            self.timeout_id = None
+            return False
+
+        return True
 
 
 class SoundConverterWindow(GladeWindow):
@@ -1078,6 +227,9 @@ class SoundConverterWindow(GladeWindow):
 
         self.set_sensitive()
         self.set_status()
+
+        # wrap the widget
+        self.progressbar = ProgressBar(self.progressbar)
 
     # This bit of code constructs a list of methods for binding to Gtk+
     # signals. This way, we don't have to maintain a list manually,
@@ -1265,17 +417,21 @@ class SoundConverterWindow(GladeWindow):
             if remaining is not None:
                 seconds = max(remaining % 60, 1)
                 minutes = remaining / 60
-                remaining = _('%d:%02d left') % (minutes, seconds)
-                self.progressbar.set_text(remaining)
+                remaining_str = _('%d:%02d left') % (minutes, seconds)
+                self.progressbar.set_text(remaining_str)
                 self.progressbar.set_show_text(True)
-                title = '{} - {}'.format(_('SoundConverter'), remaining)
+                title = '{} - {}'.format(_('SoundConverter'), remaining_str)
                 self.widget.set_title(title)
+                # -1 because the progressbar should get a chance for
+                # displaying 100% and not get hidden too early
+                fraction = duration / (duration + remaining - 1)
+                self.progressbar.set_fraction(fraction)
 
         # return True to keep the GLib timeout running
         return True
 
     def update_progress(self):
-        """Refresh all progress bars, including the total progress.
+        """Refresh all progress bars.
 
         Can be used in GLib.timeout_add.
         """
@@ -1289,9 +445,7 @@ class SoundConverterWindow(GladeWindow):
 
         if not paused and running:
             # if paused, don't refresh the progress
-            total_progress, task_progress = self.converter_queue.get_progress()
-            self.progressbar.set_fraction(total_progress)
-
+            task_progress = self.converter_queue.get_progress()[1]
             for task, progress in task_progress:
                 self.set_file_progress(task.sound_file, progress)
 
@@ -1300,6 +454,7 @@ class SoundConverterWindow(GladeWindow):
 
     def on_convert_button_clicked(self, *args):
         # reset and show progress bar
+        self.progressbar.set_fraction(0)
         self.progress_frame.show()
         self.status_frame.hide()
         self.set_status(_('Converting'))
